make: Add 'make uninstall' to uninstall corebox
After having the possibility to install corebox, it's reasonable to also
add the option to remove it again.
Doing this by hand is somewhat of a struggle as you not only have to
remove the binary, but also each link to it.

To fix this, there is now a `uninstall.sh` script, which will cleanly
remove corebox as well as the links in the specified PREFIX dir when
using `make uninstall`.